pyproject.toml
[tool.poetry]
name = "nouserbots"
version = "0.1.0"
description = "Block as-telegram bots"
authors = ["kaeeraa <kaeeraa@nebula-nook.ru>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.scripts]
noUserBots = 'noUserBots.main:run'


[tool.poetry.dependencies]
python = ">=3.12,<3.13"
hikari = "^2.1.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
uvloop = "^0.20.0"
mypy = "^1.11.2"
isort = "^5.13.2"
hikari-arc = "^1.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = "main"
disable_error_code = ["arg-type", "union-attr"]

[[tool.mypy.overrides]]
module = "vars"
disable_error_code = ["arg-type"]
